# Fixes: - Fixed crash in music game and enabled it in menu - Fixed crash in Life game - Fixed crash in invaders game when ships explode - Fixed missing sprites in invaders - Fixed mathemagic crash caused by errors in macros that do not create sprite or tile data - Fixed cyclic references in sprite and tile macros - Fixed embedding error in sprite and tile macros; they are now embedded with the code, not in a separate section - Fixed macros not being built because intermediate files were not found during the build process
# Updated the following libraries:
- rp-pico to version 0.7
- rp2040-boot2 to version 0.3
- rp2040-hal to version 0.8
- rp2040-pac to version 0.4
- embedded-graphics was updated to version 0.8.0
- usb-hid to version 0.6.1

# Updated the following libraries with code changes due to broken API:
- Replaced deprecated embedded-time library with fugit version 0.3.5 (successor) and made necessary code changes
- Updated tiled to version 0.11.1 and made necessary code changes
- Updated st7789 to version 0.7.0 and made necessary code changes

+++ games/src/blob.rs
@@ -493,8 +493,8 @@ fn make_level(level: usize) -> Vec<Wall, 64> {
                     .push(Wall {
                         bounding_box: BoundingBox::new(
                             I32x2 {
-                                x: x as i32 * WALL_SIZE,
-                                y: y as i32 * WALL_SIZE,
+                                x: x * WALL_SIZE,
+                                y: y * WALL_SIZE,
                             },
                             I32x2 {
                                 x: WALL_SIZE,
